Respect ActiveModel 'base' validations pattern
It is a common pattern for activerecord and activemodel models to
indicate errors that are not on a given field but rather on the entire
model by adding an error to the "base" attribute.  In this case the
message tends to be a complete sentence about the model state, rather
than a clause appended to the attribute name ("can't be blank").  In
this case, we don't want to append "Base" to the beginning of the error
message.